Function: Implement closures
This patch allows closures to function in the engine!
Every Function stores a `context` field which references
the execution context it was defined within. When the
functions is called, we enter that context in order to
regain access to the variables which were in scope when
that function was defined. :^)