Node: Support destructuring assignment expressions
Things are now split into two interfaces: `Declarable` and `Assignable`
`Assignable` is the interface for things that can be assigned to, ie
identifiers, destructuring patterns, and member expressions.
`Declarable` is the interface for things that can be declared, ie
variables, and destructuring patterns.

`LeftHandSideExpression` now inherits from `Assignable`, and
`AssignmentExpression` accepts only an `Assignable` as the
left-hand side.