ObjectValue: Fix long-standing bug in `set`
In f939758bfdf725f104d86f6fdefbbb923decdf74, `ObjectValue#set` was
changed to get the property descriptor to assign to via `getProperty`
This caused a bug when setting properties which exist on
an object's prototype, but not on the object itself.

The properties were being modified on the prototype, when they should
have been created on the object itself. Changing to `getOwnProperty`
fixes this as it will return `null` if the object does not contain the
specified property, causing `defineOwnProperty` to be called.