Value: Separate "ordinary functions" from "constructor functions"
Not all ECMAScript Function Objects have a [[Construct]] method.
It seems the ones which do are the ones created by a function
expression or declaration; function objects created by get/set
properties or methods are not constructors. The specification
uses OrdinaryFunctionCreate to create the latter function objects,
hence they are named "ordinary functions". Function expressions or
declarations, which yield the former function objects, begin
as ordinary functions in the specification, but are then passed
to MakeConstructor -- hence these are known as "constructor functions"